Converge `NavSection` components (#454)
