+++ src/components/Dropdown.vue
@@ -29,7 +29,7 @@
       :class="{ visible: isHovered || isFocused }"
       role="menu"
     >
-      <slot :onFocus="onFocus" />
+      <slot :onFocus="onFocus" :a11y-props="{ role: 'menuitem' }" />
     </div>
   </div>
 </template>
