+++ src/components/HeaderSection.vue
@@ -1,11 +1,6 @@
 <template>
-  <header :class="{ embedded: isEmbedded }">
-    <EmbeddedNavSection
-      v-if="isEmbedded"
-      :key="$route.path"
-      :show-nav-search="showNavSearch"
-    />
-    <NavSection v-else :key="$route.path" :show-nav-search="showNavSearch" />
+  <header class="embedded">
+    <NavSection :key="$route.path" :show-nav-search="showNavSearch" />
     <NotificationBanner v-if="showNotification" />
   </header>
 </template>
@@ -14,7 +9,6 @@
 export default {
   name: 'HeaderSection',
   props: {
-    isEmbedded: { type: Boolean, default: true },
     showNavSearch: { type: Boolean, default: false },
     showNotification: { type: Boolean, default: false },
   },
