+++ src/pages/audio/_id.vue
@@ -34,11 +34,7 @@ import { MEDIA } from '~/constants/store-modules'
 const AudioDetailPage = {
   name: 'AudioDetailPage',
   mixins: [iframeHeight],
-  layout({ store }) {
-    return store.state.nav.isEmbedded
-      ? 'embedded-with-nav-search'
-      : 'with-nav-search'
-  },
+  layout: 'with-nav-search',
   data() {
     return {
       thumbnailURL: null,
