+++ src/pages/sources.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="section" dir="ltr">
-    <div :class="['container', isEmbedded ? '' : 'is-fluid']">
+    <div class="container">
       <div class="mb-10">
         <h1 class="text-5xl mb-10">
           {{ $t('sources.title') }}
@@ -136,17 +136,13 @@
 <script>
 import sortBy from 'lodash.sortby'
 import { mapState } from 'vuex'
-import { NAV, PROVIDER } from '~/constants/store-modules'
+import { PROVIDER } from '~/constants/store-modules'
 
 const ARABIC_NUMERAL_LOCALES = ['ar', 'fa', 'ur', 'ckb', 'ps']
 
 const SourcePage = {
   name: 'source-page',
-  layout({ store }) {
-    return store.state.nav.isEmbedded
-      ? 'embedded-with-nav-search'
-      : 'with-nav-search'
-  },
+  layout: 'with-nav-search',
   data() {
     return {
       sort: {
@@ -156,7 +152,6 @@ const SourcePage = {
     }
   },
   computed: {
-    ...mapState(NAV, ['isEmbedded']),
     ...mapState(PROVIDER, ['imageProviders']),
     sortedProviders() {
       const sorted = sortBy(this.imageProviders, [this.sort.field])
